4/11/19     - Scripting is now a thing, first few bots are the usual auto-drink, an eat-all-food in inventory, a acre-clay digger     and its partner stockpile acre clay bot.         - All scripts typically will check for stamina and auto drink on their own. One difference over the old scripts         is that these will now also refill waterskins/flasks buckets from nearby barrels of water if needed     - Bot chat channel is back, if there's a prompt or event it has to wait on the user it'll be in here     - Tooltips for custom menu icons on your hotbars will now display properly     - Updated omni's sounds to the latest
4/9/19
    - Can now use pathfinding by ctrl+shift+alt clicking a tile or object. I'll improve it overtime for the few
    minor things that could use updating
    - `Show FPS` will now show your velocity

4/7/19
    - Ard fixes scent bug with flatworld
    - Softcaps calculated with Longs rather than integers..

+++ src/haven/Makewindow.java
@@ -271,7 +271,7 @@ public class Makewindow extends Widget {
             popt = opt;
         }
 
-        final List<Integer> mods = new ArrayList<>();
+        final List<Long> mods = new ArrayList<>();
         if (qmod != null) {
             g.image(qmodl.tex(), new Coord(0, qmy + 4));
             c = new Coord(xoff, qmy);
@@ -284,14 +284,14 @@ public class Makewindow extends Widget {
                         for (CharWnd.SAttr attr : ui.gui.chrwdg.skill) {
                             if (name.equals(attr.attr.nm)) {
                                 FastText.printf(g, c.add(0, t.sz().y + 2), "%d", attr.attr.comp);
-                                mods.add(attr.attr.comp);
+                                mods.add((long)attr.attr.comp);
                                 break;
                             }
                         }
                         for (CharWnd.Attr attr : ui.gui.chrwdg.base) {
                             if (name.equals(attr.attr.nm)) {
                                 FastText.printf(g, c.add(0, t.sz().y + 2), "%d", attr.attr.comp);
-                                mods.add(attr.attr.comp);
+                                mods.add((long)attr.attr.comp);
                                 break;
                             }
                         }
